# src/api_service/pkg/docs/swagger.yaml
basePath: /
definitions:
  api.CreateUrlRequest:
    properties:
      redirectUrl:
        type: string
    required:
    - redirectUrl
    type: object
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  api.UrlDto:
    properties:
      newUrl:
        type: string
      redirectUrl:
        type: string
      urlId:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a URL shortener service.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: URL Shortener API
  version: "1.0"
paths:
  /{urlId}:
    get:
      description: Redirect to url
      operationId: redirect-url
      parameters:
      - description: url id
        in: path
        name: urlId
        required: true
        type: string
      responses:
        "302":
          description: Found
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Redirect to url
      tags:
      - url
  /ping:
    get:
      description: Ping the server
      operationId: ping
      produces:
      - text/plain
      responses:
        "200":
          description: pong
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Ping the server
      tags:
      - ping
  /url:
    post:
      description: Create a new url redirect
      operationId: create-url
      parameters:
      - description: url
        in: body
        name: url
        required: true
        schema:
          $ref: '#/definitions/api.CreateUrlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UrlDto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Create a new url redirect
      tags:
      - url
  /url/{urlId}:
    get:
      description: Get url
      operationId: get-url
      parameters:
      - description: url id
        in: path
        name: urlId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UrlDto'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get url
      tags:
      - url
schemes:
- http
securityDefinitions:
  JWT_AUTH:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"